Extract bike/bici routes for origins and destinations using CycleStreets via stplanr

Extract bike/bici routes for origins and destinations using CycleStreets via stplanr

Usage

bici_routes(
  od.data,
  distance.threshold = Inf,
  plan = c("quietest", "fastest", "balanced"),
  origin.col = names(od.data)[1],
  destination.col = names(od.data)[2],
  trips.col = names(od.data)[grep("trip", names(od.data))]
)

Arguments

od.data

A sf object with OD data

distance.threshold

distance threshold in metres to create a subset of the routes

plan

one of "quietest" (default), "fastest", or "balanced" for cycle streets

origin.col

string with name of the column with origin id

destination.col

string with name of the column with destination id

trips.col

string with name of the column with number of trips for each OD pair

Value

An sf object with the routes to school

Examples

if (FALSE) { # \dontrun{
bici_routes(od_data_almada)
} # }