Origin-destination data for Almada

A dataset containing origin and destination data for the Almada region.

Format

A data frame with X rows and Y variables

Source

http://somewhere.com/

Examples

od_data_almada
#> # A tibble: 223 × 4
#> Error in `[<-`(`*tmp*`, unname, value = list(origin = c(38790L, 38795L, 38796L, 40276L, 40277L, 40283L, 40289L, 41511L, 41513L, 41516L, 41539L, 41541L, 41561L, 41564L, 41566L, 41614L, 41617L, 41618L, 41623L, 43004L, 43544L, 41733L, 41735L, 41822L, 41886L, 41888L, 41892L, 41933L, 41975L, 43263L, 43264L, 43444L, 42002L, 42794L, 43042L, 43056L, 43082L, 43097L, 43116L, 43208L, 43237L, 43249L, 43269L, 43270L, 43271L, 43276L, 43285L, 43291L, 43336L, 43357L, 43371L, 43373L, 43380L, 43382L, 43384L, 43404L, 43405L, 43407L, 43408L, 43409L, 43411L, 43413L, 43414L, 43449L, 43451L, 43463L, 43465L, 43475L, 43478L, 43487L, 43489L, 43493L, 43515L, 43516L, 43517L, 43522L, 43546L, 43547L, 43550L, 43552L, 43554L, 43602L, 43603L, 43610L, 43613L, 43614L, 43619L, 43621L, 43694L, 43732L, 43778L, 43779L, 44136L, 44452L, 43793L, 43797L, 43801L, 43803L, 43804L, 43839L, 43856L, 43862L, 43873L, 43877L, 43889L, 43890L, 43891L, 43915L, 43933L, 43937L, 43969L, 43973L, 44244L, 44248L, 43977L, 43978L, 43979L, 43980L, 43983L, 43985L, 43987L, 44008L, 44011L, 44110L, 44111L, 44137L, 44140L, 44141L, 44143L, 44146L, 44211L, 44288L, 44291L, 44301L, 44306L, 44361L, 44376L, 44429L, 44457L, 44548L, 44549L, 44556L, 44560L, 44571L, 44599L, 44624L, 44627L, 44628L, 44634L, 44713L, 44717L, 44721L, 44649L, 44651L, 44707L, 44708L, 44709L, 44852L, 44869L, 44870L, 44871L, 44878L, 44879L, 44882L, 44917L, 44918L, 44919L, 44923L, 44924L, 44937L, 44942L, 44943L, 44946L, 44947L, 44949L, 44951L, 45179L, 45182L, 45229L, 45231L, 45282L, 45284L, 45285L, 45360L, 45455L, 45504L, 45554L, 45555L, 45559L, 45561L, 45562L, 45575L, 45576L, 45578L, 45583L, 45592L, 45610L, 45744L, 45790L, 45794L, 45797L, 45801L, 45813L, 45876L, 45879L, 45881L, 45882L, 45931L, 45943L, 45965L, 45984L, 45987L, 46051L, 46054L, 46056L, 46057L, 46058L, 46103L, 46196L, 46414L, 194583L, 194590L, 201334L), destination = c(1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590, 1503590), trips = c(0.0305472767567197, 0.063394904375022, 0.117761151961283, 0.235760110014968, 0.167198016885459, 0.08792516954451, 0.0342383702342949, 0.640541216956357, 1.91979841203278, 0.653197355880584, 0.0247292796401807, 0.129423799816768, 0.0235137787131543, 13.3426207673249, 1.49365891176123, 0.0727079339632862, 0.0407265111272881, NaN, 0.0680345018623603, NaN, 0.191448402771097, 0.109402663760207, 0.176300657253518, 0.0176158810549604, 0.265296330791564, 0.922324775949329, 0.223291843263295, 1.20579948764052, NaN, 0.0397426926067649, 0.0257739451639003, 0.0695235566780299, 0.223483938809584, 0.611020694337118, 0.246110603062434, 0.267512181005983, 0.287179492263909, 0.0776004904570586, 0.0807074209974681, 0.0081178901082183, 0.529692195305144, 0.14523206539737, 0.00867354448709706, 0.0586405990871061, 0.0498087029257082, 0.0650148604729166, NaN, 0.0969163628171883, 0.418669942584009, NaN, 0.0606680099636563, 0.0882505587813493, NaN, 0.15818588196697, 0.0746662509114202, 0.287878145447163, 0.719179963890355, 0.382859795554218, NaN, 0.130176645700287, 0.134603329592391, 1.42599703366706, 0.331482586752884, 0.172072228834398, 0.845087598874331, 0.468334038128258, 0.648151245334841, 0.210079748077448, 0.0336591656188656, 0.0180280787802362, 0.0858413528236578, 0.110861062084387, 0.0733901309043899, 0.0760685930530774, 0.0209866218740327, 0.151551218564267, 0.183025355695125, 0.189612839691456, 1.43007895443769, 0.75156700249508, 0.464624712455846, 0.157978447516855, 0.302654423562226, 0.708342020868173, NaN, 0.310867051286363, 0.0109700482068581, 0.141236589450363, 0.288411081110145, 0.329707553719595, 0.0233159504799354, 0.0529116210657933, 0.0962556242633647, NaN, 0.0866061038061692, 0.0975495189176387, 0.120747179050998, NaN, 0.116406869493327, 0.933249518812667, NaN, 0.438516131867488, 1.68358675764814, 0.400713345114616, 0.0122883872244384, 0.301875699331218, 0.0258339399080008, 1.55258196484907, 0.518307218024048, 1.07859655179878, 0.116675921583832, 0.053254730403019, NaN, NaN, 0.0239973405322516, 0.00946746413100927, 0.0923311232151424, 0.0585764837885064, 0.0566316410057962, 0.0824902538229069, 0.0340716289648301, 0.0528501824304454, 0.0104044835659871, 0.321783818309246, 0.809375386191733, 0.338661752560981, 0.330860785517467, 0.100635212428937, 0.082514475666857, NaN, 0.895163253645893, 0.320602834688487, 0.490501624156908, 0.332880140839011, 0.670234836229487, NaN, NaN, 15.7880125152017, 0.107686394863679, NaN, 0.288903022804542, 0.135504509276381, 0.328285712056871, 0.196447680450638, 0.650837468234304, 0.00695951991685595, 0.153889627616892, 0.141382271134984, 0.711871063565643, 0.294244160055357, 0.29456300928151, 0.644002983661613, 0.0802172992732402, 0.435692739684825, 0.20735217471432, 0.525870568887933, 0.0110702451331615, NaN, 0.201733373573165, NaN, 0.0966384071791354, 0.967315568715285, 0.0958153294549897, 0.0994322681235525, 1.74996210924331, NaN, 0.103191869744668, 1.37050411319802, 0.918217147342635, NaN, NaN, 3.03609057433253, NaN, 0.834298579559615, 1.01250260773907, 1.70476426392267, 0.0741140079096683, 0.246037752401514, 0.912201334668799, 0.199586358326936, 2.24069592466933, 1.19794531804504, 0.100216348948616, 0.121399314744141, 0.0403579990957292, 0.090512958654321, 0.651381555428115, 0.210235076172182, 0.0192767973566765, NaN, NaN, 0.12187095873524, 0.268075962443245, 1.17858669789346, 0.0752227062853461, 0.0930166904885711, 0.207298331695715, 0.0520746351615008, 0.104642965519052, 0.20581048811356, NaN, 0.600084769913938, 0.566935350302512, 0.0283379391971516, NaN, 0.371680277945584, 0.109240746058465, 0.109392316760667, 0.192023961222718, 0.0770761974476728, 0.402156844137798, 0.270147488481031, 0.0374235450074795, 0.0455005018382574, 0.119892436188942, 1.38821876735859, NaN, NaN, 0.271162061927859, 0.38522747935736, 0.236095526749676, 0.262830634650191, 0.337293232272136), geometry = structure(list(structure(c(-9.24003754711492, -9.238283, 38.6466166646558, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23550302408749, -9.238283, 38.6469418420778, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23914105797923, -9.238283, 38.6469293683752, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22222076490582, -9.238283, 38.6478168089415, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23295787090937, -9.238283, 38.6474426140246, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23218191222138, -9.238283, 38.6486633561485, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23649740912181, -9.238283, 38.6487473474732, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23984449973227, -9.238283, 38.668538767676, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24184999713262, -9.238283, 38.6675314773921, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23952397504211, -9.238283, 38.6688578523954, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23667785661155, -9.238283, 38.6458683241879, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23823098920313, -9.238283, 38.6459395543683, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2046151390893, -9.238283, 38.665992157284, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24418267174483, -9.238283, 38.6673587476392, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24022835661988, -9.238283, 38.6680761087296, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23611766406918, -9.238283, 38.6457949603139, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23951700999594, -9.238283, 38.6460192815924, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23731381847736, -9.238283, 38.6459815741823, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23841754397647, -9.238283, 38.6463525000789, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24082081953957, -9.238283, 38.661793447602, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23369076011136, -9.238283, 38.6709918585274, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23580509167908, -9.238283, 38.6475432947321, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23756420837026, -9.238283, 38.647223848401, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23591101649206, -9.238283, 38.6432702820192, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24013263148873, -9.238283, 38.6692188591299, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24079464982842, -9.238283, 38.6692334450807, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2331011772929, -9.238283, 38.6653271434116, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23680070308744, -9.238283, 38.6669866102577, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24273601886177, -9.238283, 38.6574646773307, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23733611188158, -9.238283, 38.6439732210409, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23503191884809, -9.238283, 38.6440325118192, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22343682412541, -9.238283, 38.6566579602729, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23612756811656, -9.238283, 38.6481329202516, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23765499121511, -9.238283, 38.6562449091187, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23373853276514, -9.238283, 38.6489203577596, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24111666314657, -9.238283, 38.6454148728966, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22390482128563, -9.238283, 38.659003765842, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23896779526581, -9.238283, 38.6633320129351, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23905787022655, -9.238283, 38.6478658010874, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21964722775479, -9.238283, 38.6483230901013, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23540462064861, -9.238283, 38.6547122894169, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23394253196361, -9.238283, 38.6519319213187, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23462854731101, -9.238283, 38.6442091499704, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23364817895042, -9.238283, 38.644096109249, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2364027609007, -9.238283, 38.644239511445, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23576800200053, -9.238283, 38.6444445053823, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22593880287851, -9.238283, 38.6663831823332, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20406099224383, -9.238283, 38.6674863795233, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23698270860878, -9.238283, 38.6549636515354, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2353159129546, -9.238283, 38.6445188968566, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22040793022589, -9.238283, 38.6499619563003, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23479851153548, -9.238283, 38.6501150917871, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21479100988951, -9.238283, 38.6508186792763, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23405226566568, -9.238283, 38.6501613507684, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21807551419676, -9.238283, 38.6494153343263, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23632334143806, -9.238283, 38.6610998154125, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23107324024997, -9.238283, 38.6618959640897, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24165986779921, -9.238283, 38.6611219478782, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2353128381272, -9.238283, 38.6622649373878, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23732168891296, -9.238283, 38.661997356657, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23684442598698, -9.238283, 38.6624564163216, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23832466166998, -9.238283, 38.6600830981905, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23303608305845, -9.238283, 38.6613699220321, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24296720178926, -9.238283, 38.6638851066141, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23957530561704, -9.238283, 38.664071511033, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23627921322566, -9.238283, 38.651034867863, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23345731062514, -9.238283, 38.6510588375816, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23165052290885, -9.238283, 38.6496263763871, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23283902990923, -9.238283, 38.6498734487561, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21536070959694, -9.238283, 38.6499308898092, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23193675374553, -9.238283, 38.6501710529515, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23253400632992, -9.238283, 38.6493322226121, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23763020610705, -9.238283, 38.6575394628017, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23973219676856, -9.238283, 38.6574255505429, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22075395632183, -9.238283, 38.6574110469581, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23737753005875, -9.238283, 38.6577548998636, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23205661368721, -9.238283, 38.6711749932327, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23335357636628, -9.238283, 38.6710923632828, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23137346729606, -9.238283, 38.6709156795796, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2330126524799, -9.238283, 38.6711608510102, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23261413689917, -9.238283, 38.6711917537945, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2169142235381, -9.238283, 38.6515724332113, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22037973851196, -9.238283, 38.65237436099, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23695324729366, -9.238283, 38.652291780512, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21831271658577, -9.238283, 38.6530251282569, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23624319831448, -9.238283, 38.652505469029, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21878019788353, -9.238283, 38.65234905798, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21910176957136, -9.238283, 38.6519990181959, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23059151628861, -9.238283, 38.649681657631, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22068145731064, -9.238283, 38.6629183275842, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23571211725578, -9.238283, 38.6429984079104, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23696174977615, -9.238283, 38.6429603644806, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22286261116349, -9.238283, 38.6500890971718, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20725093595724, -9.238283, 38.6762974353993, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23759367050453, -9.238283, 38.6445844662328, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23651034514575, -9.238283, 38.6448547462379, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23555543356609, -9.238283, 38.6450889261744, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2391204903975, -9.238283, 38.6451929204692, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2380897643831, -9.238283, 38.6450500435668, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.25170089637163, -9.238283, 38.6623926858254, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24078174989707, -9.238283, 38.6648685873617, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23657118290459, -9.238283, 38.6536122683034, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23488621831001, -9.238283, 38.6557679850185, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23857381406069, -9.238283, 38.6568201595398, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23803009497331, -9.238283, 38.6455316813062, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23438452856554, -9.238283, 38.6450821855274, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23932402160017, -9.238283, 38.6456095732908, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24064901430688, -9.238283, 38.6687836106726, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24148765658776, -9.238283, 38.664041255402, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23845567683558, -9.238283, 38.6641124446473, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23453672231045, -9.238283, 38.6432693113916, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23374017785287, -9.238283, 38.6434275127646, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23463984406174, -9.238283, 38.6516704268272, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22691907614857, -9.238283, 38.6489224506143, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23497397752102, -9.238283, 38.6435455378783, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2370888500011, -9.238283, 38.6434490826624, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23317105697023, -9.238283, 38.6435498046344, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23128247067837, -9.238283, 38.6436170130312, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23437685866932, -9.238283, 38.6437638820312, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23612626305488, -9.238283, 38.6436945811084, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23549025399267, -9.238283, 38.6438767811182, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23666336695902, -9.238283, 38.646422403295, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23271438976008, -9.238283, 38.6457682997682, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23849571278967, -9.238283, 38.6551903495679, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23578861280366, -9.238283, 38.6556872257109, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23560197227031, -9.238283, 38.651309276064, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23270332351252, -9.238283, 38.651350922784, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23510016810231, -9.238283, 38.651482886285, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23681477629238, -9.238283, 38.6645908085497, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23047286625569, -9.238283, 38.6647234777922, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23625283043332, -9.238283, 38.6557683516013, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20993810187656, -9.238283, 38.6587868387244, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24176914253871, -9.238283, 38.6634414683008, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23183136998636, -9.238283, 38.6441773085241, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24167185523248, -9.238283, 38.6651645724172, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22986661024805, -9.238283, 38.6450149065576, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21184602217181, -9.238283, 38.6638052092918, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24690210087137, -9.238283, 38.6677088835915, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20790407434652, -9.238283, 38.6744723276722, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22562316859722, -9.238283, 38.6502457030341, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23574831893001, -9.238283, 38.6526842471362, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2352677914898, -9.238283, 38.6528447449277, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22201576765544, -9.238283, 38.6526605223746, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23192466475387, -9.238283, 38.6736488619018, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2295285477953, -9.238283, 38.6660758683038, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21654547752395, -9.238283, 38.6500792779107, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23556034841821, -9.238283, 38.6496985625095, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23307833463329, -9.238283, 38.6503256165564, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23732519211114, -9.238283, 38.6556144787053, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23506609099824, -9.238283, 38.6540875763749, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23689791970577, -9.238283, 38.6542330135891, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23597445755881, -9.238283, 38.6541759613411, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2162116550476, -9.238283, 38.6529202725064, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23453186888714, -9.238283, 38.6530666373995, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23765238640773, -9.238283, 38.6536360131383, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23309370197499, -9.238283, 38.6530847794878, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2114836682609, -9.238283, 38.6531545269555, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24418443405672, -9.238283, 38.6647304407653, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23334686678256, -9.238283, 38.6727717550483, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23272429453826, -9.238283, 38.6724807310645, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23234996139288, -9.238283, 38.6725632741851, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23481369002841, -9.238283, 38.6725052424712, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23175764429502, -9.238283, 38.672894988829, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23308119955628, -9.238283, 38.672642121769, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24028430066814, -9.238283, 38.6696825601155, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2322248079588, -9.238283, 38.6683200400371, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20614497802362, -9.238283, 38.6664531028138, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23906010701114, -9.238283, 38.6697401154909, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23325884033373, -9.238283, 38.670210433573, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23394019573098, -9.238283, 38.6716592761404, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23229777150639, -9.238283, 38.6717432450455, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23733292622651, -9.238283, 38.6709918698376, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23327500546056, -9.238283, 38.671803698869, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23173499652944, -9.238283, 38.671832458349, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23134105684693, -9.238283, 38.6719324150706, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23485280452169, -9.238283, 38.6719652693113, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22670851033103, -9.238283, 38.671958527058, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20510548511282, -9.238283, 38.6719071225732, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23376288225354, -9.238283, 38.6723186392893, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23337400777041, -9.238283, 38.6723761840596, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22889025732573, -9.238283, 38.6704338429753, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23954320681267, -9.238283, 38.672031198625, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20622292323863, -9.238283, 38.6749535484569, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2112911842383, -9.238283, 38.6705228859665, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23082514174528, -9.238283, 38.6598178599945, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23254529862744, -9.238283, 38.6703690450156, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23892409686553, -9.238283, 38.6576846531942, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2400868768948, -9.238283, 38.6580844370639, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21900336255482, -9.238283, 38.6581172369627, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24463855454976, -9.238283, 38.6580455612918, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24175910134053, -9.238283, 38.657750508992, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23751836287986, -9.238283, 38.6686658448345, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23186185506189, -9.238283, 38.670488976781, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23941218610452, -9.238283, 38.6587835884045, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21795413585858, -9.238283, 38.6707663483646, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20749252761573, -9.238283, 38.6765299028897, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23490507013324, -9.238283, 38.6707247693194, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20610972878212, -9.238283, 38.6765436510528, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23402299676286, -9.238283, 38.6730951368558, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23354072978212, -9.238283, 38.6732517180474, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23298937599629, -9.238283, 38.6733650824686, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23261665021089, -9.238283, 38.6734829933167, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23071928908198, -9.238283, 38.6607805217878, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23031693932866, -9.238283, 38.6556677803301, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24733419686733, -9.238283, 38.6580424071777, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22957610403858, -9.238283, 38.660497350265, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24215022536885, -9.238283, 38.6596960637871, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24233538583246, -9.238283, 38.6523613495093, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23058053634591, -9.238283, 38.659313368232, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24251190957923, -9.238283, 38.6627735890894, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24351465908468, -9.238283, 38.6586484660396, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21904494992837, -9.238283, 38.6590473264386, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22952703674124, -9.238283, 38.6592999541104, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22032550896962, -9.238283, 38.6587752252489, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22114402252861, -9.238283, 38.6587583737855, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2426211212678, -9.238283, 38.6594001654739, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23227672694809, -9.238283, 38.6595779224709, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23162506331927, -9.238283, 38.6724319371305, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24001763590609, -9.238283, 38.6698229871412, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.222343967998, -9.238283, 38.6699149804597, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23243308328513, -9.238283, 38.6447687977619, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23505526855906, -9.238283, 38.646209086159, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23183687843521, -9.238283, 38.6479748323809, 38.669886), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg"))), class = c("sfc_LINESTRING", "sfc"), precision = 0, bbox = structure(c(xmin = -9.25170089637163, ymin = 38.6429603644806, xmax = -9.20406099224383, ymax = 38.6765436510528), class = "bbox"), crs = structure(list(input = "EPSG:4326",     wkt = "GEOGCRS[\"WGS 84\",\n    ENSEMBLE[\"World Geodetic System 1984 ensemble\",\n        MEMBER[\"World Geodetic System 1984 (Transit)\"],\n        MEMBER[\"World Geodetic System 1984 (G730)\"],\n        MEMBER[\"World Geodetic System 1984 (G873)\"],\n        MEMBER[\"World Geodetic System 1984 (G1150)\"],\n        MEMBER[\"World Geodetic System 1984 (G1674)\"],\n        MEMBER[\"World Geodetic System 1984 (G1762)\"],\n        MEMBER[\"World Geodetic System 1984 (G2139)\"],\n        ELLIPSOID[\"WGS 84\",6378137,298.257223563,\n            LENGTHUNIT[\"metre\",1]],\n        ENSEMBLEACCURACY[2.0]],\n    PRIMEM[\"Greenwich\",0,\n        ANGLEUNIT[\"degree\",0.0174532925199433]],\n    CS[ellipsoidal,2],\n        AXIS[\"geodetic latitude (Lat)\",north,\n            ORDER[1],\n            ANGLEUNIT[\"degree\",0.0174532925199433]],\n        AXIS[\"geodetic longitude (Lon)\",east,\n            ORDER[2],\n            ANGLEUNIT[\"degree\",0.0174532925199433]],\n    USAGE[\n        SCOPE[\"Horizontal component of 3D system.\"],\n        AREA[\"World.\"],\n        BBOX[-90,-180,90,180]],\n    ID[\"EPSG\",4326]]"), class = "crs"), n_empty = 0L))): Assigned data `map(.subset(x, unname), vectbl_set_names, NULL)` must be
#> a vector.
#> Caused by error in `vec_size()`:
#> ! `x` must be a vector, not a <sfc_LINESTRING/sfc> object.