Origin-destination data for Almada

A dataset containing origin and destination data for the Almada region.

Format

A data frame with X rows and Y variables

Source

http://somewhere.com/

Examples

od_data_almada
#> # A tibble: 300 × 4
#> Error in `[<-`(`*tmp*`, unname, value = list(origin = c(38790L, 38794L, 38795L, 38796L, 40276L, 40277L, 40281L, 40283L, 40285L, 40289L, 41539L, 41541L, 41614L, 41617L, 41618L, 41619L, 41623L, 41706L, 43004L, 41729L, 41732L, 41733L, 41735L, 41788L, 41792L, 41793L, 41795L, 41796L, 41797L, 41798L, 41800L, 41801L, 41817L, 41818L, 41821L, 41822L, 41823L, 41882L, 41883L, 41892L, 41933L, 41975L, 43261L, 43263L, 43264L, 43444L, 42001L, 42002L, 42139L, 42140L, 42492L, 42500L, 42505L, 42794L, 43039L, 43040L, 43041L, 43042L, 43054L, 43055L, 43056L, 43082L, 43097L, 43105L, 43110L, 43112L, 43113L, 43116L, 43122L, 43123L, 43124L, 43135L, 43208L, 43222L, 43224L, 43227L, 43230L, 43232L, 43237L, 43243L, 43244L, 43249L, 43265L, 43266L, 43268L, 43269L, 43270L, 43271L, 43272L, 43274L, 43276L, 43285L, 43336L, 43357L, 43362L, 43363L, 43364L, 43366L, 43371L, 43373L, 43380L, 43382L, 43383L, 43384L, 43385L, 43394L, 43400L, 43401L, 43403L, 43404L, 43405L, 43407L, 43408L, 43409L, 43411L, 43413L, 43414L, 43449L, 43451L, 43463L, 43465L, 43475L, 43477L, 43478L, 43482L, 43487L, 43489L, 43493L, 43515L, 43516L, 43517L, 43522L, 43602L, 43603L, 43607L, 43610L, 43613L, 43614L, 43619L, 43621L, 43628L, 43641L, 43642L, 43657L, 43658L, 43659L, 43694L, 43732L, 43744L, 43747L, 43754L, 43755L, 43762L, 43763L, 43765L, 43767L, 43768L, 43772L, 43776L, 43778L, 43779L, 43865L, 43867L, 44131L, 44132L, 44133L, 44136L, 43785L, 43786L, 43793L, 43794L, 43797L, 43798L, 43799L, 43801L, 43803L, 43804L, 43839L, 43856L, 43858L, 43859L, 43860L, 43862L, 43873L, 43877L, 43888L, 43889L, 43890L, 43891L, 43933L, 43937L, 43969L, 43970L, 43973L, 43974L, 43976L, 44243L, 44244L, 44246L, 44248L, 44520L, 43977L, 43978L, 43979L, 43980L, 43983L, 43985L, 43987L, 44008L, 44011L, 44079L, 44080L, 44081L, 44084L, 44092L, 44094L, 44096L, 44097L, 44100L, 44102L, 44104L, 44105L, 44110L, 44111L, 44113L, 44122L, 44123L, 44125L, 44137L, 44140L, 44141L, 44143L, 44146L, 44171L, 44175L, 44177L, 44179L, 44211L, 44288L, 44291L, 44301L, 44306L, 44360L, 44361L, 44507L, 44511L, 44523L, 44527L, 44536L, 44540L, 44548L, 44549L, 44556L, 44560L, 44599L, 44623L, 44624L, 44627L, 44628L, 44634L, 44713L, 44717L, 44721L, 44649L, 44651L, 44707L, 44708L, 44709L, 44741L, 44746L, 44748L, 44750L, 44852L, 45455L, 45554L, 45555L, 45559L, 45561L, 45562L, 45578L, 45813L, 45876L, 45879L, 45881L, 45882L, 45931L, 45943L, 45965L, 45984L, 45987L, 46051L, 46054L, 46056L, 46057L, 46058L, 194583L, 194589L, 194590L, 200686L, 201334L), destination = c(1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836, 1503836), trips = c(0.301019106533819, 0.351298893579093, 0.695200116691603, 1.13654723983106, 1.87551872005329, 1.70589134126942, NaN, 0.689512380343769, 0.276465610415217, 0.246387932898931, 0.330242913029768, 1.59433443656501, 1.00616492257135, 0.463500504723761, NaN, 0.0104872331219928, 0.763420201350796, 0.0321331593841332, NaN, 1.10581576314327, NaN, 1.04344115813707, 1.70617524879423, 0.097478113329825, 6.06182227309174, NaN, 0.418448542936115, NaN, 9.18554424232839, 0.674339918773624, 4.61976190287156, 0.169434287919284, 0.0563847969242611, 0.0149303187998694, 1.29906742958238, 0.420544020663893, 0.234220691497594, 0.0627363026164425, 0.164585970052983, 0.0489376546215615, 0.156456355582252, NaN, 0.0789245738002872, 0.762281834900765, 0.54141955527131, 0.114881033895156, 2.51637052020516, 1.85766548745541, 0.189231692010704, 0.235025573585759, 0.181667661059738, 0.195238510211547, 0.171183237621937, 0.825727007094876, 0.048872553440577, 0.26787957237121, 0.273233730891116, 1.80007214800708, 0.0583327705518057, 0.297785858244845, 3.12695431905349, 0.306968335076309, 0.0215859368667286, 0.674103181349253, 0.251439838722296, 0.304903125675757, 0.644170691761484, 0.644339408142724, 1.07511122998951, 0.916574130794989, 0.93353512349357, 0.0862210212342611, 0.0559477623571102, 3.46407727772927, 1.53242335979215, 0.353050984998067, 1.25838502387098, 2.21385003293178, 1.04924168574871, 0.0700457605748533, 0.202058573350713, 0.54534966893552, 0.114455854222857, 0.0669003841302837, 0.202816645648751, 0.178112611822952, 1.24529808885004, 0.949584874404349, 0.0427207714438295, 1.84447054483066, 1.21858292081271, NaN, 0.760247311897191, NaN, 0.128356888587831, 0.0692185443778139, 0.2155316630922, 0.0348041119223438, 0.32656843322459, 0.487722634503496, NaN, 0.876059221903881, 0.0719317670576068, 0.421416922061998, 0.0365501552709828, 0.093490573049792, 0.0938064234310819, 0.324758815323853, 0.169044698075408, 0.1368943525935, 0.342994089602218, 0.17083986823636, NaN, 0.0497379397048196, 0.0469299884946341, 0.817805195987597, 0.163591731740651, 0.0437516780286429, 0.199301844378896, 2.05064954370285, 2.97093978539137, 1.33959036027469, NaN, 0.201644872861607, 0.594697525789352, 0.0901745277018648, 0.485575376527156, 0.749706404759249, 0.0745999682569545, 0.0765843499307782, 0.0318912611365797, 0.14758265031546, 0.631911256214663, 1.08850920646532, 0.0562547229426418, 2.31741553492643, NaN, 0.985290568914636, 0.0398627611943598, 0.537008657459509, 5.71860667348689, 1.21687809650695, 0.0651492364727287, 1.95570008855958, 0.466405997374289, 2.75655315421927, 1.82204862074161, 0.206194277904771, 0.105310189658404, 0.0880801208220812, 2.96514856460229, 0.370271680008181, NaN, 2.2852484974633, 0.669080385043469, 0.671174942862663, 1.44836794948724, 0.389267224296137, 0.152077033121293, 0.642725355159958, 1.26319327346645, 0.121577913904929, 0.0553799911652632, 1.63387781476529, 0.638648180822599, 0.912425082226276, 0.525523485696993, 0.130817267734485, 0.172205760167172, 1.44994378635532, 0.442481314249475, 1.62660625321408, 1.0283561213341, 0.180134993581796, 1.98309718085927, NaN, 1.72943490911426, 0.384500097694532, NaN, 0.190931540582319, 0.161850726533847, 0.0898175826272516, 1.08159409116751, 2.67012369416579, 0.469247467526751, 0.397635373699601, 0.166115580130787, 5.10872468080131, 0.321880682900172, 0.124506682066053, 0.253778834242961, 2.93636272150532, 1.29571096959412, 1.31015058959096, 0.0320333084829176, 0.381377606954294, 0.0299947932484249, NaN, NaN, NaN, 0.100155203622467, 0.561534942077536, 0.198358307707132, 2.21198339430914, 1.33192075068211, 1.28255113204728, 1.7846745106027, 0.728937427070769, 0.627035228770395, 0.153310502825502, 3.05421539090635, 4.49816547887125, NaN, 0.732901446488619, NaN, 0.0466681092666871, 1.12147002156089, 1.0762945817872, 0.853807354563674, 13.6902142530756, 0.17522497098262, 1.74495892303628, 0.539416255969904, 1.28315908385657, 0.149472393912829, 0.206294066954906, 0.0548349259956952, 0.0378974585326391, 1.53419673841586, 1.43633515148443, 0.408921672214441, 0.0180078612709081, NaN, 0.30468661258736, 1.45514798049038, 2.82103098708938, 2.10358668833005, 1.38109752114119, 0.47800217285558, 0.134076645620425, 6.84639987641251, 0.127930487316015, 0.213918410168976, NaN, 0.183012533494294, 0.342032021147716, 1.13745717073427, 0.0249270131759418, 0.200229981518489, NaN, NaN, 0.889223562017239, 0.43874068880813, 1.12904893688342, 0.155179786991899, 0.879924857630826, 0.0346821576729859, 0.918660605954512, 0.765031358747129, 1.11237477265498, 0.673389656077105, 0.629402428364684, 1.41980709429668, 0.260326857904029, 1.26133750461002, 0.497284381420482, 1.55364108578932, 0.0342471272747601, 0.286349534164391, 0.160114271960326, 0.0423327191295566, 0.0104163610224369, NaN, 0.0296771757942743, 0.627361154830784, 0.182493034330894, 0.0269717609585161, NaN, NaN, 0.88599299040614, 0.343515047600699, 0.0502621216880946, NaN, 0.248542363247637, 0.0659436602800032, 0.307658056290127, 0.157919275779229, 0.0244196352773465, 0.300988699359483, 0.325486020461476, 0.0318881364045976, 0.0555529454347092, 0.143887996530741, 0.890758606046044, NaN, 4.31780484862921, 0.340098569409383, 3.4222902045555, 0.506207183699713, 3.07432245867055), geometry = structure(list(structure(c(-9.24003754711492, -9.233949, 38.6466166646558, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20408736635629, -9.233949, 38.6464973444334, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23550302408749, -9.233949, 38.6469418420778, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23914105797923, -9.233949, 38.6469293683752, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22222076490609, -9.233949, 38.6478168089426, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23295787090846, -9.233949, 38.6474426140267, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20807213679695, -9.233949, 38.6486892852168, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23218191220831, -9.233949, 38.6486633561433, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20507965079385, -9.233949, 38.6482244839695, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23649740912181, -9.233949, 38.6487473474732, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23667785661155, -9.233949, 38.6458683241879, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23823098920313, -9.233949, 38.6459395543683, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23611766406918, -9.233949, 38.6457949603139, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23951700999594, -9.233949, 38.6460192815924, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23731381847736, -9.233949, 38.6459815741823, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20689637557082, -9.233949, 38.6463231993294, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23841754397647, -9.233949, 38.6463525000789, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21537255092234, -9.233949, 38.6272128730971, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24082081953957, -9.233949, 38.661793447602, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23488205651924, -9.233949, 38.6414984760547, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20823444352876, -9.233949, 38.6471383063576, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23580509167908, -9.233949, 38.6475432947321, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23756420837026, -9.233949, 38.647223848401, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21603594451517, -9.233949, 38.6396878466843, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23458850819284, -9.233949, 38.6396621699505, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20569815725901, -9.233949, 38.6406231009823, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23393920912596, -9.233949, 38.6416631484451, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20670521858599, -9.233949, 38.6395726935879, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23021313681299, -9.233949, 38.6402838074097, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20824043753377, -9.233949, 38.6414885210235, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23342086890536, -9.233949, 38.6414143755611, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20970769192353, -9.233949, 38.6419790049545, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20306629094927, -9.233949, 38.6424485297555, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2080021361825, -9.233949, 38.64246662537, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23222776932681, -9.233949, 38.6430912110099, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23591101649206, -9.233949, 38.6432702820192, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20507983222434, -9.233949, 38.642815014918, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20543302370836, -9.233949, 38.6447058210379, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20882220035821, -9.233949, 38.644024379781, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23310117729813, -9.233949, 38.665327143412, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2368007030868, -9.233949, 38.6669866102594, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24273601886177, -9.233949, 38.6574646773307, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21047188890249, -9.233949, 38.6439624481028, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23733611188158, -9.233949, 38.6439732210409, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23503191884809, -9.233949, 38.6440325118192, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.223436824125, -9.233949, 38.656657960273, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22597459236702, -9.233949, 38.6442711747433, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23612756811656, -9.233949, 38.6481329202516, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21918609603588, -9.233949, 38.6222596270224, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20567549540606, -9.233949, 38.6258984030445, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21555029386774, -9.233949, 38.6251112480919, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21702114809641, -9.233949, 38.6270812511895, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21896549743751, -9.233949, 38.6470635544872, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23765499121511, -9.233949, 38.6562449091187, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20512081303962, -9.233949, 38.630275553705, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22049827962299, -9.233949, 38.6297904267436, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2067521643252, -9.233949, 38.6305107507966, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23373853276514, -9.233949, 38.6489203577596, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20641239903203, -9.233949, 38.6468638592749, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20570293006055, -9.233949, 38.6460519627554, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24111666314657, -9.233949, 38.6454148728966, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2239048212861, -9.233949, 38.6590037658419, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23896779526607, -9.233949, 38.6633320129343, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20683748347738, -9.233949, 38.6273452364933, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20510794329847, -9.233949, 38.6283794205402, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21857148881083, -9.233949, 38.6282533254383, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20316202906077, -9.233949, 38.6286892559681, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23905787023075, -9.233949, 38.6478658010878, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22076905297579, -9.233949, 38.6460569334804, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2026202926393, -9.233949, 38.6427551438805, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22284608358041, -9.233949, 38.6452877484086, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22212713615835, -9.233949, 38.6316728257061, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21964722775496, -9.233949, 38.6483230901014, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22955429752726, -9.233949, 38.635169692192, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22516659433711, -9.233949, 38.6294057986331, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22971912632357, -9.233949, 38.6356189216019, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22997647691005, -9.233949, 38.6359050600979, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23072648483816, -9.233949, 38.6357538186967, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23540462064861, -9.233949, 38.6547122894169, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21984600654801, -9.233949, 38.6322716406806, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20483842274376, -9.233949, 38.6332469441308, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23394253196361, -9.233949, 38.6519319213187, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20773166776994, -9.233949, 38.6443374303574, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21121721244318, -9.233949, 38.6442146658705, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20665469707992, -9.233949, 38.6439874778108, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23462854731101, -9.233949, 38.6442091499704, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23364817895042, -9.233949, 38.644096109249, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2364027609007, -9.233949, 38.644239511445, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20851211745053, -9.233949, 38.644560878815, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23522726053153, -9.233949, 38.639030175954, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23576800200204, -9.233949, 38.64444450537, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2259388028787, -9.233949, 38.6663831823333, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23698270860878, -9.233949, 38.6549636515354, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23531591295241, -9.233949, 38.6445188968394, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21328689267871, -9.233949, 38.6411286920781, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20392158051797, -9.233949, 38.644521618494, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20819774787792, -9.233949, 38.6448127813014, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21604009001969, -9.233949, 38.6488405779579, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22040793022589, -9.233949, 38.6499619563003, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23479851153548, -9.233949, 38.6501150917871, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21479100988951, -9.233949, 38.6508186792763, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23405226566568, -9.233949, 38.6501613507684, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20399968740791, -9.233949, 38.649816490207, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21807551419676, -9.233949, 38.6494153343263, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20826641845855, -9.233949, 38.6504573468211, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21741578758699, -9.233949, 38.6295924594168, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20422537975949, -9.233949, 38.6309884066762, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2090129411882, -9.233949, 38.6294727005612, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20289234564408, -9.233949, 38.6304718062904, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23632334142852, -9.233949, 38.6610998154174, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23107324024997, -9.233949, 38.6618959640897, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24165986779921, -9.233949, 38.6611219478782, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23531283812889, -9.233949, 38.662264937395, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23732168891481, -9.233949, 38.6619973566458, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23684442598517, -9.233949, 38.6624564163326, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23832466167226, -9.233949, 38.6600830981893, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23303608305837, -9.233949, 38.6613699220322, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24296720178926, -9.233949, 38.6638851066141, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23957530561479, -9.233949, 38.6640715110322, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23627921322566, -9.233949, 38.651034867863, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23345731062514, -9.233949, 38.6510588375816, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23165052290885, -9.233949, 38.6496263763871, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20120969126696, -9.233949, 38.6483875145777, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23283902990923, -9.233949, 38.6498734487561, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21274609688354, -9.233949, 38.6472314964763, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21536070959694, -9.233949, 38.6499308898092, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23193675374553, -9.233949, 38.6501710529515, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23253400632992, -9.233949, 38.6493322226121, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23763020596669, -9.233949, 38.6575394630912, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23973219676856, -9.233949, 38.6574255505429, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22075395632185, -9.233949, 38.657411046958, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23737753005875, -9.233949, 38.6577548998636, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2169142235381, -9.233949, 38.6515724332113, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22037973851196, -9.233949, 38.65237436099, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20761602820834, -9.233949, 38.6520283796028, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23695324729456, -9.233949, 38.6522917805137, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21831271658577, -9.233949, 38.6530251282569, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23624319831448, -9.233949, 38.652505469029, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21878019788353, -9.233949, 38.65234905798, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21910176957136, -9.233949, 38.6519990181959, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23073058404784, -9.233949, 38.6428029486844, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22685055556881, -9.233949, 38.6384257220042, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20408289394341, -9.233949, 38.6384302494763, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23479252148118, -9.233949, 38.6418822522536, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23561528668745, -9.233949, 38.6418154944184, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23655480922011, -9.233949, 38.6418123702646, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23059151629043, -9.233949, 38.649681657629, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22068145731059, -9.233949, 38.6629183275842, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20854080827367, -9.233949, 38.6400418982716, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20869683949441, -9.233949, 38.6406061525821, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23242148527384, -9.233949, 38.6419851780964, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23100640783209, -9.233949, 38.6422651147311, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23573363303948, -9.233949, 38.6426407392307, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23477403572858, -9.233949, 38.6427086937237, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23668193618558, -9.233949, 38.6424644100711, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21061494319768, -9.233949, 38.6396543233567, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23406075480624, -9.233949, 38.6427526915505, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23354820517869, -9.233949, 38.6427766237411, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20923368376699, -9.233949, 38.6431300996127, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23571211725578, -9.233949, 38.6429984079104, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23696174978462, -9.233949, 38.6429603644766, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20371161843445, -9.233949, 38.6318188721191, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20532936216505, -9.233949, 38.6318979116618, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.233643696061, -9.233949, 38.6420437987103, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23442659282185, -9.233949, 38.6422817385279, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23575081152623, -9.233949, 38.6422827870141, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22286261116348, -9.233949, 38.6500890971718, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21466705507185, -9.233949, 38.6295416905792, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20309810334236, -9.233949, 38.6338095952988, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23759367050453, -9.233949, 38.6445844662328, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21613550594582, -9.233949, 38.6439437941363, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23651034515425, -9.233949, 38.644854746235, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21801040460458, -9.233949, 38.6385552388602, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20748528367866, -9.233949, 38.6450354479072, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23555543355779, -9.233949, 38.6450889261847, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2391204903975, -9.233949, 38.6451929204692, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2380897643831, -9.233949, 38.6450500435668, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.25170089637164, -9.233949, 38.6623926858254, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24078174990161, -9.233949, 38.6648685873664, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23288854247654, -9.233949, 38.6428897872013, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21730765657924, -9.233949, 38.6422545447152, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21086817019791, -9.233949, 38.6414378454267, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23657118290459, -9.233949, 38.6536122683034, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23488621829087, -9.233949, 38.6557679850207, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23857381412829, -9.233949, 38.6568201594954, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20871926460301, -9.233949, 38.6455027762133, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23803009497331, -9.233949, 38.6455316813062, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23438452856554, -9.233949, 38.6450821855274, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23932402160017, -9.233949, 38.6456095732908, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24148765658787, -9.233949, 38.6640412554019, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23845567683536, -9.233949, 38.6641124446473, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23453672231045, -9.233949, 38.6432693113916, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22797822300826, -9.233949, 38.6415591562221, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23374017785287, -9.233949, 38.6434275127646, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20341639056113, -9.233949, 38.6431212405786, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2051734083065, -9.233949, 38.6437272805339, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20987688181732, -9.233949, 38.6516720371122, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23463984406174, -9.233949, 38.6516704268272, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21141740821793, -9.233949, 38.6504884119076, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22691907614844, -9.233949, 38.6489224506142, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20673748177544, -9.233949, 38.634134074729, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23497397752102, -9.233949, 38.6435455378783, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2370888500011, -9.233949, 38.6434490826624, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23317105697023, -9.233949, 38.6435498046344, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23128247067837, -9.233949, 38.6436170130312, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23437685866932, -9.233949, 38.6437638820312, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23612626305488, -9.233949, 38.6436945811084, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23549025399267, -9.233949, 38.6438767811182, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23666336695902, -9.233949, 38.646422403295, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23271438976008, -9.233949, 38.6457682997682, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22844041444198, -9.233949, 38.6379490192681, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22407201977866, -9.233949, 38.6350081383986, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20816761347648, -9.233949, 38.637129910552, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20123167419537, -9.233949, 38.6362183226684, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23407998924971, -9.233949, 38.640618745414, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20753023511059, -9.233949, 38.6398398426745, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22123813213508, -9.233949, 38.6375661648537, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23472631921174, -9.233949, 38.6410280910515, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2348309479304, -9.233949, 38.6412298253165, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2320813910258, -9.233949, 38.6400778233528, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20878583355412, -9.233949, 38.6411268568316, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2342661137424, -9.233949, 38.6401617844834, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23849571278967, -9.233949, 38.6551903495679, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23578861280359, -9.233949, 38.6556872257127, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20391322211457, -9.233949, 38.6536112194402, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20559035975596, -9.233949, 38.6349900047287, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20403645489336, -9.233949, 38.6352957161002, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2040704421503, -9.233949, 38.6417237832312, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23560197227031, -9.233949, 38.651309276064, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23270332351252, -9.233949, 38.651350922784, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23510016810231, -9.233949, 38.651482886285, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23681477629216, -9.233949, 38.6645908085488, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2304728662329, -9.233949, 38.6647234777906, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21426230200852, -9.233949, 38.6367934685214, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22964350028787, -9.233949, 38.6385792861686, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23063563832421, -9.233949, 38.6377368363387, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2288991217936, -9.233949, 38.6388024668713, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23625283043224, -9.233949, 38.6557683516003, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20993810187653, -9.233949, 38.6587868387244, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24176914253837, -9.233949, 38.6634414683009, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23183136998636, -9.233949, 38.6441773085241, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24167185523248, -9.233949, 38.6651645724172, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20079610815017, -9.233949, 38.6441143199152, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22986661024786, -9.233949, 38.6450149065577, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23022630437255, -9.233949, 38.6362146584334, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22893130393693, -9.233949, 38.6361131586721, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22805746849418, -9.233949, 38.6366863249497, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21133624659976, -9.233949, 38.6334263582242, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20730615366016, -9.233949, 38.6463281211422, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20844507196003, -9.233949, 38.6393853525914, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22562316859729, -9.233949, 38.6502457030339, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23574831893001, -9.233949, 38.6526842471362, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2352677914898, -9.233949, 38.6528447449277, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22201576765525, -9.233949, 38.6526605223748, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22952854780082, -9.233949, 38.6660758683024, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22918817747926, -9.233949, 38.6377195574179, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21654547752395, -9.233949, 38.6500792779107, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23556034841821, -9.233949, 38.6496985625095, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23307833463329, -9.233949, 38.6503256165564, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23732519211114, -9.233949, 38.6556144787053, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23506609099824, -9.233949, 38.6540875763749, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23689791970577, -9.233949, 38.6542330135891, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23597445755881, -9.233949, 38.6541759613411, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21621165504699, -9.233949, 38.6529202725066, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23453186888714, -9.233949, 38.6530666373995, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23765238640664, -9.233949, 38.6536360131362, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2330937019747, -9.233949, 38.653084779488, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2114836682604, -9.233949, 38.6531545269553, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20584423797701, -9.233949, 38.6289859264045, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20692176286879, -9.233949, 38.6296835969201, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20408674454427, -9.233949, 38.6296378066433, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.20612283521523, -9.233949, 38.6296913899052, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24418443405674, -9.233949, 38.6647304407654, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23082514174542, -9.233949, 38.6598178599928, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23892409686553, -9.233949, 38.6576846531942, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24008687689897, -9.233949, 38.6580844370673, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21900336254826, -9.233949, 38.6581172369717, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24463855454827, -9.233949, 38.6580455612919, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24175910133279, -9.233949, 38.6577505089951, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23941218610452, -9.233949, 38.6587835884045, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23071928908193, -9.233949, 38.6607805217885, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2303169393288, -9.233949, 38.6556677803301, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24733419686734, -9.233949, 38.6580424071777, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22957610403927, -9.233949, 38.6604973502638, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24215022536848, -9.233949, 38.6596960637867, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24233538583241, -9.233949, 38.6523613495093, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23058053634591, -9.233949, 38.659313368232, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24251190957923, -9.233949, 38.6627735890894, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.24351465908472, -9.233949, 38.6586484660388, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2190449499304, -9.233949, 38.6590473264359, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22952703673979, -9.233949, 38.6592999541128, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22032550896962, -9.233949, 38.6587752252489, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.22114402252411, -9.233949, 38.6587583737859, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2426211212678, -9.233949, 38.6594001654739, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23227672694809, -9.233949, 38.6595779224709, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23243308328513, -9.233949, 38.6447687977619, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.2097993709525, -9.233949, 38.6435065387666, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23505526855906, -9.233949, 38.646209086159, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.21763657221886, -9.233949, 38.645476181518, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg")), structure(c(-9.23183687843205, -9.233949, 38.6479748323896, 38.640666), dim = c(2L, 2L), class = c("XY", "LINESTRING", "sfg"))), class = c("sfc_LINESTRING", "sfc"), precision = 0, bbox = structure(c(xmin = -9.25170089637164, ymin = 38.6222596270224, xmax = -9.20079610815017, ymax = 38.6669866102594), class = "bbox"), crs = structure(list(input = "EPSG:4326",     wkt = "GEOGCRS[\"WGS 84\",\n    ENSEMBLE[\"World Geodetic System 1984 ensemble\",\n        MEMBER[\"World Geodetic System 1984 (Transit)\"],\n        MEMBER[\"World Geodetic System 1984 (G730)\"],\n        MEMBER[\"World Geodetic System 1984 (G873)\"],\n        MEMBER[\"World Geodetic System 1984 (G1150)\"],\n        MEMBER[\"World Geodetic System 1984 (G1674)\"],\n        MEMBER[\"World Geodetic System 1984 (G1762)\"],\n        MEMBER[\"World Geodetic System 1984 (G2139)\"],\n        MEMBER[\"World Geodetic System 1984 (G2296)\"],\n        ELLIPSOID[\"WGS 84\",6378137,298.257223563,\n            LENGTHUNIT[\"metre\",1]],\n        ENSEMBLEACCURACY[2.0]],\n    PRIMEM[\"Greenwich\",0,\n        ANGLEUNIT[\"degree\",0.0174532925199433]],\n    CS[ellipsoidal,2],\n        AXIS[\"geodetic latitude (Lat)\",north,\n            ORDER[1],\n            ANGLEUNIT[\"degree\",0.0174532925199433]],\n        AXIS[\"geodetic longitude (Lon)\",east,\n            ORDER[2],\n            ANGLEUNIT[\"degree\",0.0174532925199433]],\n    USAGE[\n        SCOPE[\"Horizontal component of 3D system.\"],\n        AREA[\"World.\"],\n        BBOX[-90,-180,90,180]],\n    ID[\"EPSG\",4326]]"), class = "crs"), n_empty = 0L))): Assigned data `map(.subset(x, unname), vectbl_set_names, NULL)` must be
#> a vector.
#> Caused by error in `vec_size()`:
#> ! `x` must be a vector, not a <sfc_LINESTRING/sfc> object.